Move objects and case classes to from package object to CypherType file
